+++ picnic_app/lib/pages/vote/vote_detail_achieve_page.dart
@@ -418,7 +418,7 @@ class _VoteDetailAchievePageState extends ConsumerState<VoteDetailAchievePage> {
                   width: width,
                   child: PicnicCachedNetworkImage(
                     imageUrl: voteModel.mainImage!,
-                    width: width.toInt(),
+                    width: width,
                     memCacheWidth: width.toInt(),
                   ),
                 ),
